fix support for MediaWiki 1.26 to fix a transaction exception
For some reasons in 1.26 (and maybe other versions) the extension
DeleteBatch is not able to call $db->startAtomic() without then
calling $wikipage->doDeleteArticle() and causing an exception
since a transaction is already opened.

In my honest opinion this should not happen since the call
to $wikipage->doDeleteArticle is done with the argoment $commit
parameter set to false, so it should not open another transaction
internally to then try to commit it.

Sadly I do not understand 90% of the workflow so as workaround
I've just disabled the $db->startAtomic() in version <= 1.26.